refactor: layout meta
