+++ apps/dashboard/app/lib/root-links.ts
@@ -1,0 +1,3 @@
+import type { Route } from "../../.react-router/types/app/+types/root";
+
+export const links: Route.LinksFunction = () => [];
