+++ apps/dashboard/app/lib/root-meta.ts
@@ -1,0 +1,9 @@
+import type { Route } from "../../.react-router/types/app/+types/root";
+
+export const meta: Route.MetaFunction = () => [
+  { charset: "utf-8" },
+  { title: "Panelium" },
+  { name: "description", content: "E" },
+  { name: "viewport", content: "width=device-width, initial-scale=1" },
+  { name: "theme-color", content: "#ffffff" },
+];
